Consertando problema de kill(pid, SIGSTOP) para um pid NULO, porque ao iniciarmos o scheduler, o currentProcess->procPID era nulo... Isso estava causando uma segfault. Agora, parece que de fato estamos executando os programas, mas obtemos agora um loop infinito

+++ scheduler.c
@@ -16,7 +16,7 @@ const int numPriorities = 8;
 const float quantum = 2;
 
 typedef struct {
-    char* name;
+    char name[30];
     int type, priority, start, duration, finished;
     pid_t procPid;
 } Process;
@@ -156,7 +156,6 @@ void newProcessHandler(int signal)
     printf("params[0]: %d\n", params[0]);
 
     strcpy(p.name, programName);
-    //p.name = "p1";
     p.type = params[0];
     
     // Initializes process and stops execution in order to obtain pid
@@ -167,25 +166,26 @@ void newProcessHandler(int signal)
     } else {
         execve(programName, NULL, NULL);
     }
-    
-    if(params[0] == 1) {
-        newRoundRobin(p);
-    }
-    else if(params[0] == 2) {
-        p.priority = params[1];
-        newPriority(p);
-    }
-    else if(params[0] == 3) {
-        p.start = params[1];
-        p.duration = params[2];
-        newRealTime(p);
-    }
-    else return;// invalid value params[0]
-    
-    printf("%s -> program read\n", programName);
-    
+    if(pid != 0) { 
+        if(params[0] == 1) {
+            newRoundRobin(p);
+        }
+        else if(params[0] == 2) {
+            p.priority = params[1];
+            newPriority(p);
+        }
+        else if(params[0] == 3) {
+            p.start = params[1];
+            p.duration = params[2];
+            newRealTime(p);
+        }
+        else return;// invalid value params[0]
+        
+        printf("%s -> program read\n", programName);
+    } 
     shmdt(programName);
     shmdt(params);
+
 }
 
 void scheduler() {
@@ -200,19 +200,16 @@ void scheduler() {
     //
     //sig_t t = signal(SIGUSR1, newProcessHandler);
     //printf("num te falei %d\n", signal(SIGUSR1, newProcessHandler) );
-    
     if (signal(SIGUSR1, newProcessHandler) < 0) 
     {
         printf("error registering signal\n");
     }
     printf("registered signal handler\n");
-    
+   
     initProcessesQueues();
-    
     clock_t stTime, curTime;
     stTime = clock();
     int currentSecond = 0;
-    
     Process* curProcess = NULL;
     int result, status; // temporary variables to store process meta-data
     
@@ -222,11 +219,9 @@ void scheduler() {
     // 1 if a round robin process is in execution, 0 otherwise
     int executingRoundRobinProcess = 0;
     clock_t rrStartTime;
-    
     while(1) {
         curTime = clock();
         currentSecond = ((curTime - stTime) / CLOCKS_PER_SEC) % 60;
-        
         if (executingRealTimeProcess) {
             result = waitpid(curProcess->procPid, &status, WNOHANG);
             if (result != 0) { // process finished execution
@@ -258,7 +253,6 @@ void scheduler() {
         // check which process to execute next. Next process will
         if (!executingRealTimeProcess) {
             Process* nProcess = nextProcess();
-            
             if (nProcess != NULL) { // there is an enqueued process to be executed
                 // if a round robin process is executing we must check if its time
                 // is up before switching processes in case the next process has
@@ -274,7 +268,7 @@ void scheduler() {
                         kill(curProcess->procPid, SIGCONT); // resume real time process
                     }
                 } else {
-                    kill(curProcess->procPid, SIGSTOP); // stop current process
+                    if(curProcess != NULL) kill(curProcess->procPid, SIGSTOP); // stop current process
                     curProcess = nProcess;
                     kill(curProcess->procPid, SIGCONT); // resume real time process
                 }
@@ -299,6 +293,7 @@ void scheduler() {
         // falta tambem indicar que os processos acabaram, para que nosso scheduler eventualmente morra e tal...
         // vamos ter que usar algo do tipo wait(pid_processo, &status, ALGUMA FLAG)...
         //
+        sleep(1);
     }
     
     shmdt(scheduler_pid);
